refactor(core): remove hotSwapFiberType from the public API
It was a playground-only mutation utility that leaked into core in a
"fix" commit; core stays read-only instrumentation. The swap logic now
lives inline in the vite-playground source editor, its sole consumer.
